refactor: move Url.public to View.public
